host/requirements.txt
pyserial >= 3
pycryptodome >= 3.9.4
opencv-contrib-python >= 4.1.2.30
numpy >= 1.18.1
psutil >= 5.6.7
sense-hat >= 2.2.0